Refactor: Put index.html and main js file at top level
Allows JS file to read solutions.csv
