+++ index.html
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <head>
-  <link type="text/css" rel="stylesheet" href="./puzz.css">
+  <link type="text/css" rel="stylesheet" href="./webapp/puzz.css">
 </head>
 <body>
   <div class="displayContainer centerFlex">
